refactor(gridsat): rename metadata module
Rename metadata module to netcdf_metadata. Update import statements to
use netcdf_metadata module.
