+++ src/goesdl/gridsat/netcdf_metadata.py
